Added attributes to a linked img and a file img
Added height and width to the linked TOP image in the index page, as
well as alt text. Added a file image to the about page. This is
finishing the links and images lesson.

This is also doing the "Commit Messages" assignment. I'm using
VS code to write the commit message after inputting git commit in the
terminal.

This marks the end of the HTML section right up before the "Recipes"
project.